requirements.txt
anthropic==0.25.6
faiss-cpu==1.8.0
google-generativeai==0.5.2
langchain==0.1.16
openai==1.23.6
pdf2image==1.17.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
streamlit==1.33.0
tiktoken==0.6.0
PyPDF2==3.0.1
pdfminer.six
pillow-heif
opencv-python
unstructured[all-docs]
